Fixed issue where the change from BCELoss to BCEWithLogitsLoss had caused a regression in the test.py script, namely that we now needed to apply a sigmoid to the model output before evaluating as label membership probabilities.
Fixed an issue where the test.py script was applying a filter of label probability >0.5 to determine the predicted labels, but this was not being applied in the train.py script, leading to very low (and incorrect) Jaccard indices in training logs (e.g. 0.07) with reasonable and accurate ones in test.py output.